app/requirements.txt
flask
psycopg2-binary
gunicornflask==3.0.3
psycopg2-binary==2.9.9
gunicorn==23.0.0
azure-identity==1.13.0
azure-keyvault-secrets==4.4.0